test: e2e 테스트 분리
도메인별로 e2e테스트를 분리함.
